+++ app/com/mohiva/play/silhouette/impl/providers/oauth2/GitHubProvider.scala
@@ -19,7 +19,6 @@
  */
 package com.mohiva.play.silhouette.impl.providers.oauth2
 
-import com.mohiva.play.silhouette._
 import com.mohiva.play.silhouette.api.LoginInfo
 import com.mohiva.play.silhouette.api.util.HTTPLayer
 import com.mohiva.play.silhouette.impl.exceptions.ProfileRetrievalException
@@ -44,7 +43,7 @@ abstract class GitHubProvider(httpLayer: HTTPLayer, stateProvider: OAuth2StatePr
   extends OAuth2Provider(httpLayer, stateProvider, settings) {
 
   /**
-   * The content type returned from the provider.
+   * The content type to parse a profile from.
    */
   type Content = JsValue
 
@@ -84,28 +83,47 @@ abstract class GitHubProvider(httpLayer: HTTPLayer, stateProvider: OAuth2StatePr
           val docURL = (json \ "documentation_url").asOpt[String]
 
           throw new ProfileRetrievalException(SpecifiedProfileError.format(id, msg, docURL))
-        case _ => Future.from(parseProfile(parser, json))
+        case _ => profileParser.parse(json)
       }
     }
   }
+}
+
+/**
+ * The profile parser for the common social profile.
+ */
+class GitHubProfileParser extends SocialProfileParser[JsValue, CommonSocialProfile] {
 
   /**
-   * Defines the parser which parses the most common profile supported by Silhouette.
+   * Parses the social profile.
    *
-   * @return The parser which parses the most common profile supported by Silhouette.
+   * @param json The content returned from the provider.
+   * @return The social profile from given result.
    */
-  protected def parser: Parser = (json: JsValue) => {
+  def parse(json: JsValue) = Future.successful {
     val userID = (json \ "id").as[Long]
     val fullName = (json \ "name").asOpt[String]
     val avatarUrl = (json \ "avatar_url").asOpt[String]
     val email = (json \ "email").asOpt[String].filter(!_.isEmpty)
 
     CommonSocialProfile(
-      loginInfo = LoginInfo(id, userID.toString),
+      loginInfo = LoginInfo(ID, userID.toString),
       fullName = fullName,
       avatarURL = avatarUrl,
       email = email)
   }
+}
+
+/**
+ * The profile builder for the common social profile.
+ */
+trait GitHubProfileBuilder extends CommonSocialProfileBuilder {
+  self: GitHubProvider =>
+
+  /**
+   * The profile parser implementation.
+   */
+  val profileParser = new GitHubProfileParser
 }
 
 /**
@@ -133,6 +151,6 @@ object GitHubProvider {
    * @return An instance of this provider.
    */
   def apply(httpLayer: HTTPLayer, stateProvider: OAuth2StateProvider, settings: OAuth2Settings) = {
-    new GitHubProvider(httpLayer, stateProvider, settings) with CommonSocialProfileBuilder
+    new GitHubProvider(httpLayer, stateProvider, settings) with GitHubProfileBuilder
   }
 }